+++ packages/express/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 2.1.21
+
+### Patch Changes
+
+- Updated dependencies [[`a036ce8`](https://github.com/clerk/javascript/commit/a036ce8fef3b3ee2b49fd05d592b083ffc37f463)]:
+  - @clerk/shared@4.13.1
+  - @clerk/backend@3.4.13
+
 ## 2.1.20
 
 ### Patch Changes