+++ packages/tanstack-react-start/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @clerk/tanstack-react-start
 
+## 1.3.1
+
+### Patch Changes
+
+- Updated dependencies [[`a036ce8`](https://github.com/clerk/javascript/commit/a036ce8fef3b3ee2b49fd05d592b083ffc37f463)]:
+  - @clerk/shared@4.13.1
+  - @clerk/react@6.7.1
+  - @clerk/backend@3.4.13
+
 ## 1.3.0
 
 ### Minor Changes