+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @clerk/vue
 
+## 2.3.1
+
+### Patch Changes
+
+- Updated dependencies [[`a036ce8`](https://github.com/clerk/javascript/commit/a036ce8fef3b3ee2b49fd05d592b083ffc37f463)]:
+  - @clerk/shared@4.13.1
+
 ## 2.3.0
 
 ### Minor Changes